requirements.txt
streamlit
transformers
pdfminer.six
torch
